Update functions file structure according to new convention
